requirements.txt
ansible
docopt
jinja2
markdown
pyembed
pyembed-markdown
python-dateutil
pytz
pyyaml
